fix code at line 117 that splits string into an array
Splitting the string stored in the expression variable at each
character results in an array in which each substring refers to
a button press on the calculator.
The new variable expressionArray could therefore serve a input history
of button presses on the calculator.
The associated psudocode and comments have also been rewritten.